+++ packages/sleuth_mcp/test/flutter_daemon/daemon_session_test.dart
@@ -778,7 +778,7 @@ class _FlakyConnectBridge extends FakeVmBridge {
             'connectionMode': 'basic',
             'schemaVersion': 1,
             'sessionUuid': 'u',
-            'data': {'packageVersion': '0.35.0'},
+            'data': {'packageVersion': '0.36.0'},
           },
         });
 
@@ -873,7 +873,7 @@ class _AllDeadBridge extends FakeVmBridge {
             'connectionMode': 'basic',
             'schemaVersion': 1,
             'sessionUuid': 'u',
-            'data': {'packageVersion': '0.35.0'},
+            'data': {'packageVersion': '0.36.0'},
           },
         });
 
@@ -892,7 +892,7 @@ class _TimeoutThenLiveBridge extends FakeVmBridge {
             'connectionMode': 'basic',
             'schemaVersion': 1,
             'sessionUuid': 'u',
-            'data': {'packageVersion': '0.35.0'},
+            'data': {'packageVersion': '0.36.0'},
           },
         });